backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
yfinance
pandas
pydantic